src/requirements.txt
-illow==8.3.1
absl-py==1.0.0
attrs==21.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.24
cython-bbox==0.1.3
cytoolz==0.11.0
dask==2021.10.0
dataclasses==0.6
DCNv2==0.1
easydict==1.9
efficientnet-pytorch==0.7.1
efficientunet-pytorch==0.0.6
et-xmlfile==1.1.0
ffmpeg==1.4
flake8==3.9.2
flake8-import-order==0.18.1
fonttools==4.25.0
fsspec==2022.1.0
future==0.18.2
fvcore==0.1.5.post20210924
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.11.0
iniconfig==1.1.1
iopath==0.1.9
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.3.2
kornia==0.5.0
lap==0.4.0
locket==0.2.1
lpips==0.1.4
Markdown==3.3.6
matplotlib==3.4.3
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
motmetrics==1.2.0
munkres==1.1.4
nested-dict==1.61
networkx==2.6.3
numpy==1.20.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.3.56
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
partd==1.2.0
Pillow==9.0.1
pip==21.0.1
pluggy==1.0.0
portalocker==2.3.2
progress==1.6
protobuf==3.18.0
py==1.10.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==3.0.4
pytest==6.2.5
pytest-benchmark==3.4.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.7.1
setuptools==58.0.4
sip==4.19.13
six==1.16.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.2.9
toml==0.10.2
toolz==0.11.2
torch==1.7.0
torchfile==0.1.0
torchgeometry==0.1.2
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.8.0
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
visdom==0.1.8.9
websocket-client==1.2.3
Werkzeug==2.0.3
wheel==0.37.0
xmltodict==0.12.0
yacs==0.1.8
zipp==3.7.0
